[py] Fixed a major bug with chatrooms.
The bug was as fallows.
Let us have 3 agents: A, B, S.
S listens, A and B connect to it.
What is observed is that S reads all well but when it transmits it's either to A OR to B in turn.

The reason for this behaviour is that handle_input() uses input() which blocks.
Thus anything the server types is sent to only one of it's clients.

The solution is to send to eveuone obviously.
The implemented solution should work, but is a proof of concept, unfinished, broken.
This branch is supposed to bring the task to completion - to allow one-to-many communication.