[py] Significant refactoring to sock.py to adhere to the new protocol; the new test fixed ... and red.
Firstly, specificts of the protocol were considered.
There is no reason to reverse the message.
So: Send chunk 0, then chunk 1 until the last one and then signature.

Array index 0 is to be regarded as the oldest received or waiting to be sent chunk.
New work is added to the tail, work to be committed is pulled from the head.